components 분리
